Refactor gcs module into parent dir
