+++ src/version-compare/api.js
@@ -1,5 +1,5 @@
 const {fileMetadata: md} = require("../db/api.js");
-const gcs = require("./gcs.js");
+const gcs = require("../gcs.js");
 
 module.exports = {
   compare(newEntry) {
